chore(sources): Refactor `source_sender` into modules
